[tech debt] Reduce complexity in repos and improve boundaries
With the current set of changes we introduce an additional AreaRepository to provide the relevant data related only to Area bounded meals.
The main purpose of this change was to reduce complexity in CategoriesRepository which was having too much responsibilities.

We have also introduced sub-packages in the domain layer to improve organization and readibility.
As a result of this, we have indicated a redundant logic which was cleaned.
