+++ data/src/main/java/app/cookery/repositories/categories/CategoriesRepository.kt
@@ -9,27 +9,14 @@ class CategoriesRepository @Inject constructor(
     private val store: CategoriesStore
 ) {
 
-    fun observeCategoryDetailsByName(categoryName: String) = store.observeCategoryDetailsByName(categoryName)
-    fun observeCategoryDetailsByArea(area: String) = store.observeCategoryDetailsByArea(area)
     fun observeRandomCategoryDetails() = store.observeRandomCategoryMeals()
-    fun observeRandomAreaMeals() = store.observeRandomAreaMeals()
     fun observeAllMealCategories() = store.observeAllMealCategories()
     fun observeCategoryWithCategoryDetails(categoryName: String) =
         store.observeCategoryWithCategoryDetails(categoryName)
 
-    fun observeAreaWithCategoryDetails(areaName: String) =
-        store.observeAreaWithCategoryDetails(areaName)
-
-    fun observeAreaMeals() = store.observeAreaMeals()
-
     suspend fun fetchAllMealCategories() {
         val response = dataSource.getAllMealCategories().getOrThrow()
         store.saveAllMealCategories(response)
-    }
-
-    suspend fun fetchAllMealAreas() {
-        val response = dataSource.getMealAreas().getOrThrow()
-        store.saveAreaMeals(response)
     }
 
     suspend fun fetchMealsByCategory(categoryName: String) {
@@ -38,15 +25,6 @@ class CategoriesRepository @Inject constructor(
             meals = response,
             categoryName = categoryName,
             area = ""
-        )
-    }
-
-    suspend fun fetchMealsByArea(area: String) {
-        val response = dataSource.getMealsByArea(area).getOrThrow()
-        store.saveCategoryDetails(
-            meals = response,
-            categoryName = "",
-            area = area
         )
     }
 }
