+++ domain/src/main/java/app/cookery/domain/interactors/InitializeHomeScreenData.kt
@@ -3,6 +3,7 @@ package app.cookery.domain.interactors
 import app.cookery.AppCoroutineDispatchers
 import app.cookery.domain.Interactor
 import app.cookery.repositories.categories.CategoriesRepository
+import app.cookery.repositories.categories.areas.AreaRepository
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
@@ -10,13 +11,14 @@ private val categoryTypes = listOf("Pasta", "Starter", "Dessert")
 private val areaTypes = listOf("Italian", "Thai", "Mexican", "American")
 
 class InitializeHomeScreenData @Inject constructor(
-    private val repository: CategoriesRepository,
+    private val categoriesRepository: CategoriesRepository,
+    private val areasRepository: AreaRepository,
     private val dispatchers: AppCoroutineDispatchers
 ) : Interactor<InitializeHomeScreenData.Params>() {
     override suspend fun doWork(params: Params) {
         withContext(dispatchers.io) {
-            repository.fetchAllMealCategories()
-            repository.fetchAllMealAreas()
+            categoriesRepository.fetchAllMealCategories()
+            areasRepository.fetchAllMealAreas()
             fetchInitialCategories()
             fetchInitialAreas()
         }
@@ -24,13 +26,13 @@ class InitializeHomeScreenData @Inject constructor(
 
     private suspend fun fetchInitialCategories() {
         for (category in categoryTypes) {
-            repository.fetchMealsByCategory(category)
+            categoriesRepository.fetchMealsByCategory(category)
         }
     }
 
     private suspend fun fetchInitialAreas() {
         for (area in areaTypes) {
-            repository.fetchMealsByArea(area)
+            areasRepository.fetchMealsByArea(area)
         }
     }
