+++ domain/src/main/java/app/cookery/domain/interactors/areas/UpdateAreas.kt
@@ -1,13 +1,13 @@
-package app.cookery.domain.interactors
+package app.cookery.domain.interactors.areas
 
 import app.cookery.AppCoroutineDispatchers
 import app.cookery.domain.Interactor
-import app.cookery.repositories.categories.CategoriesRepository
+import app.cookery.repositories.categories.areas.AreaRepository
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 class UpdateAreas @Inject constructor(
-    private val repository: CategoriesRepository,
+    private val repository: AreaRepository,
     private val dispatchers: AppCoroutineDispatchers
 ) : Interactor<UpdateAreas.Params>() {
