+++ domain/src/main/java/app/cookery/domain/observers/areas/ObserveAreas.kt
@@ -1,13 +1,13 @@
-package app.cookery.domain.observers
+package app.cookery.domain.observers.areas
 
 import app.cookery.data.entities.categories.Area
 import app.cookery.domain.SubjectInteractor
-import app.cookery.repositories.categories.CategoriesRepository
+import app.cookery.repositories.categories.areas.AreaRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
 
 class ObserveAreas @Inject constructor(
-    private val repository: CategoriesRepository
+    private val repository: AreaRepository
 ) : SubjectInteractor<ObserveAreas.Params, List<Area>>() {
 
     override fun createObservable(params: Params): Flow<List<Area>> {
