+++ domain/src/main/java/app/cookery/domain/observers/categories/ObserveCategories.kt
@@ -1,4 +1,4 @@
-package app.cookery.domain.observers
+package app.cookery.domain.observers.categories
 
 import app.cookery.data.entities.categories.Category
 import app.cookery.domain.SubjectInteractor
@@ -10,7 +10,7 @@ class ObserveCategories @Inject constructor(
     private val repository: CategoriesRepository
 ) : SubjectInteractor<ObserveCategories.Params, List<Category>>() {
 
-    override fun createObservable(params: ObserveCategories.Params): Flow<List<Category>> {
+    override fun createObservable(params: Params): Flow<List<Category>> {
         return repository.observeAllMealCategories()
     }
